Add a websocket client to user namespace
I wanted to be able to test the websocket connection in the project,
so I created some code for that purpose in the user namespace.

The connect function will attempt to connect to a running server.
The send-msg function will take any clojure type, encode with transit,
and then send to the server.
The close function will close the connection.

Also updated core namespace to use config.edn.

The thing of most interest is that info.sunng/ring-jetty9-adapter
and stylefruits/gniazdo depend on different versions of jetty. This
was causing me a massive headache wrt trying to run the server or
the client. The :exclusions listed in the :dev profile of project.clj
are what solved that issue, I also put a link there for the future
in case I forget.